Refact

+++ scripts/uisorescue.sh
@@ -3,10 +3,9 @@
 source /etc/instance.conf
 SUBVOL="@archiso"
 ISO="archlinux-x86_64.iso"
-MIRROR="https://mirror.ams1.nl.leaseweb.net/archlinux/iso/latest/"
+MIRROR="mirror.ams1.nl.leaseweb.net"
 FOLDER="/mnt/archiso"
 
-ALLDONE=0
 ROOTDRIVE=$(mount | grep -Po '^.*(?= on \/ type btrfs)')
 
 [ "$ROOTDRIVE" == "" ] && echo "This script works only with BTRFS" && exit 1
@@ -22,49 +21,32 @@ function writetogrub(){
 }
 
 
-function mountarchiso(){
-	sudo mkdir -p $FOLDER && sudo mount -o compress=zstd:3,subvol=$SUBVOL $ROOTDRIVE $FOLDER
-}
-
-function umountarchiso(){
-	cd / && sudo umount $FOLDER && sudo rm -r $FOLDER
-}
-
-
 function checkiso(){
-	cd $FOLDER && curl -s $MIRROR"sha256sums.txt" | grep $ISO | sha256sum -c --
+	curl -s "https://"$MIRROR"/archlinux/iso/latest/sha256sums.txt" | grep $ISO | sha256sum -c --
 }
 
 
-#If it already done on btrfs disk?
-if [ "$(sudo btrfs subvolume list / | grep 'top level [0-9] path '$SUBVOL)" != "" ]
+if [ "$(sudo btrfs subvolume list / | grep 'top level [0-9] path '$SUBVOL)" == "" ]
 then
-    writetogrub
-    mountarchiso
-    checkiso && echo "The system already has latest iso image of Archlinux. Nothing to do" && ALLDONE=1
-    umountarchiso
-    [ $ALLDONE -eq 1 ] && exit 1
+	sudo mount $ROOTDRIVE /mnt
+	cd /mnt && sudo btrfs subvolume create $SUBVOL && cd /
+	sudo umount /mnt && writetogrub || echo "Error of creating $SUBVOL!"
 fi
 
-cd /home/$(whoami)
-echo "Downloading $ISO from " $(echo $MIRROR | cut -d"/" -f3) "..."
-curl -L -O -C - "$MIRROR$ISO"
-curl -s $MIRROR"sha256sums.txt" | grep $ISO | sha256sum -c --
 
-if [ $? -eq 0 ]
+if [ "$(sudo btrfs subvolume list / | grep 'top level [0-9] path '$SUBVOL)" != "" ]
 then
-    # make subvolume if it isn't exist and update grub entry:
-    sudo mount $ROOTDRIVE /mnt && cd /mnt && sudo btrfs subvolume create $SUBVOL && cd / && sudo umount /mnt
-    [ $? -eq 0 ] || { echo "Error of creating $SUBVOL"; exit 1; }
-    writetogrub
-    mountarchiso
-    if [ $? -eq 0 ]
+    sudo mkdir -p $FOLDER && sudo mount -o compress=zstd:3,subvol=$SUBVOL $ROOTDRIVE $FOLDER
+    cd $FOLDER
+    if checkiso
     then
-        sudo curl -o "$FOLDER/$ISO" "FILE:///home/$(whoami)/$ISO"
-        checkiso || echo "Checksum error!"
-        umountarchiso
-        rm /home/$(whoami)/$ISO
+        echo "The system already has latest iso image of Archlinux. Nothing to do"
     else
-        echo "Error of mounting $SUBVOL"
+        echo "Downloading $ISO from $MIRROR ..."
+        cd /home/$(whoami)
+        curl -L -O -C - "https://"$MIRROR"/archlinux/iso/latest/$ISO"
+        sudo curl -o "$FOLDER/$ISO" "FILE:///home/$(whoami)/$ISO"
+        checkiso && rm /home/$(whoami)/$ISO || echo "Checksum error!"
     fi
+    cd / && sudo umount $FOLDER && sudo rm -r $FOLDER
 fi
